pyproject.toml
[project]
name = "paper-intelligence"
version = "0.1.0"
description = "MCP server for intelligent paper/PDF management with RAG capabilities"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Strand AI", email = "hello@strand.ai" }
]
keywords = ["mcp", "pdf", "rag", "embeddings", "paper", "research", "llm", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Indexing",
]
dependencies = [
    "mcp[cli]>=1.0.0",
    "marker-pdf>=1.0.0",
    "chromadb>=0.4.0",
    "llama-index-core>=0.10.0",
    "llama-index-embeddings-huggingface>=0.2.0",
    "llama-index-vector-stores-chroma>=0.1.0",
    "pydantic>=2.0.0",
    "torch>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/Strand-AI/paper-intelligence"
Repository = "https://github.com/Strand-AI/paper-intelligence"
Issues = "https://github.com/Strand-AI/paper-intelligence/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "build>=1.0.0",
    "twine>=4.0.0",
]

[project.scripts]
paper-intelligence = "paper_intelligence.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["paper_intelligence"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
